+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Default maximum age of an entry in the beacon cache changed from 1h 45m to 45m
 - Removed AppMon functionality
 - Removed Support for Java 6
+- Maximum length of reported error/crash stacktrace has been limited to 128k.
+- Maximum length of reported error/crash reason has been limited to 1000.
 
 ## 2.2.0 [Release date: 2021-04-22]
 [GitHub Releases](https://github.com/Dynatrace/openkit-java/releases/tag/v2.2.0)